extracted command logic to single service LeadsPurger, to call it in PurgeCommand + CronjobListener
