+++ src/EventListener/CronjobListener.php
@@ -4,19 +4,15 @@ namespace Terminal42\LeadsBundle\EventListener;
 
 use Contao\CoreBundle\Monolog\ContaoContext;
 use Psr\Log\LoggerInterface;
-use Psr\Log\LogLevel;
-use Symfony\Bundle\FrameworkBundle\Console\Application;
-use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Output\BufferedOutput;
-use Symfony\Component\HttpKernel\KernelInterface;
+use Terminal42\LeadsBundle\Service\LeadsPurger;
 
 class CronjobListener
 {
 
     /**
-     * @var KernelInterface
+     * @var LeadsPurger
      */
-    private $kernel;
+    private $purger;
 
     /**
      * @var LoggerInterface
@@ -25,32 +21,24 @@ class CronjobListener
 
     /**
      * CronjobListener constructor.
-     * @param KernelInterface $kernel
+     * @param LeadsPurger $purger
      * @param LoggerInterface $logger
      */
-    public function __construct(KernelInterface $kernel, LoggerInterface $logger)
+    public function __construct(LeadsPurger $purger, LoggerInterface $logger)
     {
-        $this->kernel = $kernel;
+        $this->purger = $purger;
         $this->logger = $logger;
     }
 
     function onDaily()
     {
-
         try {
-            $application = new Application($this->kernel);
-            $input = new ArrayInput(
-                [
-                    'command' => 'leads:purge',
-                ]
-            );
-            $output = new BufferedOutput();
-            $application->run($input, $output);
-
+            $this->purger->execute();
         } catch (\Exception $exception) {
-            $logLevel = LogLevel::ERROR;
-            $logMessage = $exception->getMessage();
-            $this->logger->log($logLevel, $logMessage, array('contao' => new ContaoContext(__METHOD__, $logLevel)));
+            $this->logger->error(
+                $exception->getMessage(),
+                ['contao' => new ContaoContext(__METHOD__, ContaoContext::ERROR)]
+            );
         }
 
     }
