+++ src/Resources/config/listener.yml
@@ -38,7 +38,7 @@ services:
     terminal42_leads.listener.cronjob:
         class: Terminal42\LeadsBundle\EventListener\CronjobListener
         arguments:
-            - "@kernel"
+            - "@terminal42_leads.purger"
             - "@logger"
         tags:
             - { name: 'monolog.logger', channel: 'terminal42_leads' }
